send list of users as file resource in 2fa script, and make it common for user monitoring
